requirements.txt
accelerate==1.0.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==2.4.1
attrs==25.1.0
beautifulsoup4==4.13.3
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
clean-text==0.6.0
click==8.1.7
cloudpathlib==0.19.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
convokit==3.0.0
cryptography==44.0.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
emoji==1.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl#sha256=f72abb34bdf174876bd4267b29b2501677e605e0a251fdc56c163003182ed68b
executing==2.1.0
faiss-cpu==1.10.0
filelock==3.16.1
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.2.3
gensim==4.3.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.25.1
idna==3.8
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kagglehub==0.3.4
kiwisolver==1.4.7
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-text-splitters==0.3.5
langcodes==3.4.0
langsmith==0.3.4
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
msgpack-numpy==0.4.8
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
openai==1.61.0
opencv-python==4.10.0.84
orjson==3.10.15
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.2
preshed==3.0.9
prompt_toolkit==3.0.47
propcache==0.2.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.9.1
pydantic-settings==2.7.1
pydantic_core==2.23.3
Pygments==2.18.0
pymongo==4.9.1
pyparsing==3.1.4
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==3.4.1
setuptools==74.1.2
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.6
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
srsly==2.4.8
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.29
ultralytics-thop==2.0.11
Unidecode==1.3.8
urllib3==2.2.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.16.0
yarl==1.18.3
zstandard==0.23.0
